requirements.txt
altair==5.1.2
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
deepl==1.21.0
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.11.0
filelock==3.17.0
fonttools==4.54.1
fsspec==2025.2.0
gitdb==4.0.11
GitPython==3.1.40
graphviz==0.20.3
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.33.1
idna==3.4
importlib-metadata==6.8.0
ipython==8.27.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.4.2
jsonpickle==3.3.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.7
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2
numpy==1.26.1
openai==1.91.0
openpyxl==3.1.4
packaging==23.2
pandas==2.1.1
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==4.24.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==13.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.8.1b0
Pygments==2.16.1
pygount==2.0.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-docx==1.1.2
pytz==2023.3.post1
pyvis==0.3.2
PyYAML==6.0.2
rdflib==7.0.0
referencing==0.30.2
regex==2024.11.6
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
serpapi==0.1.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.39.0
streamlit-agraph==0.0.45
sympy==1.14.0
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.0
torch==2.7.1
tornado==6.3.3
tqdm==4.66.5
traitlets==5.14.3
transformers==4.53.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wcwidth==0.2.13
zipp==3.17.0
